agent/requirements.txt
langchain-docling
langchain-huggingface
langchain-community
langchain-ollama
qdrant-client[fastembed]
langgraph-supervisor
langchain-google-genai
langchain-groq
tavily-python
python-dotenv
rich
requests
sentence-transformers
mem0ai
pymongo
langgraph-checkpoint-mongodb
minio
fastapi
uvicorn[standard]
python-multipart